refactor: merging layouts to pages
